+++ CHANGELOG.md
@@ -6,6 +6,8 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- Adds `appSettings` prop to blocks that have their settings exposed on `__RUNTIME__`.
 
 ## [8.104.3] - 2020-06-09
 ### Fixed